Restructure code to build the module
