+++ .github/workflows/build-master.yml
@@ -1,0 +1,26 @@
+name: Build master branch
+
+on:
+    push:
+        branches:
+            - master
+
+jobs:
+    build:
+
+        runs-on: ubuntu-latest
+        steps:
+            - uses: actions/checkout@v2
+            - name: Set up JDK 1.8
+              uses: actions/setup-java@v1
+              with:
+                  java-version: 1.8
+            - name: Grant execute permission for gradlew
+              run: chmod +x gradlew
+            - name: Build with Gradle
+              env:
+                  GITHUB_TOKEN: ${{ secrets.BALLERINA_STDLIB_TOKEN }}
+                  packageUser: ${{ github.actor }}
+                  packagePAT: ${{ secrets.BALLERINA_STDLIB_TOKEN }}
+              run: |
+                  ./gradlew publish
