+++ .github/workflows/daily-build.yml
@@ -1,0 +1,39 @@
+name: Daily build
+
+on:
+    schedule:
+        - cron:  '0 */12 * * *'
+
+jobs:
+    ubuntu-build:
+
+        runs-on: ubuntu-latest
+        steps:
+            - uses: actions/checkout@v1
+            - name: Set up JDK 1.8
+              uses: actions/setup-java@v1
+              with:
+                  java-version: 1.8
+            - name: Build with Gradle
+              env:
+                  GITHUB_TOKEN: ${{ secrets.BALLERINA_STDLIB_TOKEN }}
+                  packageUser: ${{ github.actor }}
+                  packagePAT: ${{ secrets.BALLERINA_STDLIB_TOKEN }}
+              run: ./gradlew build
+
+    windows-build:
+
+        runs-on: windows-latest
+
+        steps:
+            - uses: actions/checkout@v1
+            - name: Set up JDK 1.8
+              uses: actions/setup-java@v1
+              with:
+                  java-version: 1.8
+            - name: Build with Gradle
+              env:
+                  GITHUB_TOKEN: ${{ secrets.BALLERINA_STDLIB_TOKEN }}
+                  packageUser: ${{ github.actor }}
+                  packagePAT: ${{ secrets.BALLERINA_STDLIB_TOKEN }}
+              run: ./gradlew.bat build
