+++ .github/workflows/publish-release.yml
@@ -1,0 +1,28 @@
+name: Publish release
+
+on:
+    release:
+        types: [published]
+
+jobs:
+    publish-release:
+        runs-on: ubuntu-latest
+        steps:
+            - uses: actions/checkout@v2
+            - name: Set up JDK 1.8
+              uses: actions/setup-java@v1
+              with:
+                  java-version: 1.8
+            - name: Grant execute permission for gradlew
+              run: chmod +x gradlew
+            - name: Publish artifact
+              env:
+                  GITHUB_TOKEN: ${{ secrets.BALLERINA_STDLIB_TOKEN }}
+                  BALLERINA_CENTRAL_ACCESS_TOKEN: ${{ secrets.BALLERINA_CENTRAL_ACCESS_TOKEN }}
+                  packageUser: ${{ github.actor }}
+                  packagePAT: ${{ secrets.BALLERINA_STDLIB_TOKEN }}
+              run: |
+                  NEW_VERSION=$(echo "${GITHUB_REF}" | cut -d "/" -f3)
+                  echo "New version: ${NEW_VERSION}"
+                  echo "Github username: ${GITHUB_ACTOR}"
+                  ./gradlew -PpublishToCentral=true -Pversion=${NEW_VERSION} publish
