--- config-api/build.gradle
@@ -1,81 +1,0 @@
-/*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-apply from: "$rootDir/gradle/balNativeLibProject.gradle"
-apply from: "$rootDir/gradle/baseNativeStdLibProject.gradle"
-
-configurations.testCompileClasspath {
-    resolutionStrategy {
-        preferProjectModules()
-    }
-}
-
-dependencies {
-    baloCreat project(':lib-creator')
-
-    implementation project(':ballerina-lang')
-    implementation project(':ballerina-config')
-    implementation project(':ballerina-runtime')
-    implementation project(':ballerina-tool')
-    implementation project(':ballerina-stringutils')
-
-    baloImplementation project(path: ':ballerina-io', configuration: 'baloImplementation')
-    baloImplementation project(path: ':ballerina-system', configuration: 'baloImplementation')
-    baloImplementation project(path: ':ballerina-lang:annotations', configuration: 'baloImplementation')
-    baloImplementation project(path: ':ballerina-runtime-api', configuration: 'baloImplementation')
-    baloImplementation project(path: ':ballerina-cache', configuration: 'baloImplementation')
-    baloImplementation project(path: ':ballerina-time', configuration: 'baloImplementation')
-    baloImplementation project(path: ':ballerina-stringutils', configuration: 'baloImplementation')
-    baloImplementation project(path: ':ballerina-java', configuration: 'baloImplementation')
-
-    interopImports project(':ballerina-io')
-    interopImports project(':ballerina-system')
-
-    testCompile project(':ballerina-tool')
-    testCompile project(':ballerina-lang:annotations')
-    testCompile project(':ballerina-runtime-api')
-    testCompile project(':ballerina-system')
-    testCompile project(':ballerina-core')
-    testCompile project(':ballerina-time')
-    testCompile project(':ballerina-reflect')
-    testCompile project(':ballerina-file')
-    testCompile project(path: ':ballerina-test-utils', configuration: 'shadow')
-    testCompile 'org.slf4j:slf4j-jdk14'
-    testCompile 'org.testng:testng'
-}
-
-createBalo {
-    jvmTarget = 'true'
-}
-
-description = 'Ballerina - Config API'
-
-test {
-    environment "testEnvironmentVariable", "EnvironmentVariable"
-    environment "user_name", "b7auser"
-    environment "password", "b7apw"
-    environment "hello_http_host", "192.168.1.11"
-    environment "hello_http_port", "5656"
-    environment "hello_cache_enabled", "true"
-    environment "hello_eviction_fac", "0.2333333"
-    environment "helloneg_http_port", "invalid port"
-    environment "helloneg_eviction_fac", "invalid eviction factor"
-}
-
-configurations.all {
-    resolutionStrategy.preferProjectModules()
-}
