+++ config-api/src/test/java/org/ballerinalang/stdlib/config/ConfigTest.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.ballerinalang.stdlib.config;
+package org.ballerinalang.stdlib.configtest;
 
 import org.ballerinalang.config.ConfigProcessor;
 import org.ballerinalang.config.ConfigRegistry;
