+++ config-api/src/test/java/org/ballerinalang/stdlib/config/EnvVarConfigTest.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.stdlib.config;
+package org.ballerinalang.stdlib.configtest;
 
 import org.ballerinalang.config.ConfigRegistry;
 import org.ballerinalang.model.values.BBoolean;
