+++ config-ballerina/Ballerina.toml
@@ -7,7 +7,7 @@ target = "java8"
 
     [[platform.libraries]]
     artifactId = "config"
-    version = "1.0.0"
-    path = "./lib/ballerina-config-1.0.0-java.jar"
+    version = "1.0.0-SNAPSHOT"
+    path = "../config-native/build/libs/config-native-1.0.0-SNAPSHOT.jar"
     groupId = "ballerina"
     modules = ["config-api"]
