+++ config-ballerina/build.gradle
@@ -1,0 +1,206 @@
+/*
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import org.apache.tools.ant.taskdefs.condition.Os
+
+description = 'Ballerina - Config Ballerina Generator'
+
+configurations {
+    jbalTools
+    configBallerinaBirDir
+    configBallerinaJar
+    externalJars
+}
+
+dependencies {
+    jbalTools ("org.ballerinalang:jballerina-tools:${ballerinaLangVersion}") {
+        transitive = false
+    }
+    compile project(':config-native')
+}
+
+clean {
+    delete "$project.projectDir/target"
+    delete "$project.projectDir/Ballerina.lock"
+}
+
+jar {
+    manifest {
+        attributes('Implementation-Title': project.name,
+                'Implementation-Version': project.version)
+    }
+}
+
+task unpackJballerinaTools(type: Copy) {
+    configurations.jbalTools.resolvedConfiguration.resolvedArtifacts.each { artifact ->
+        from zipTree(artifact.getFile())
+        into new File("${buildDir}/target/extracted-distributions", "jballerina-tools-zip")
+    }
+}
+
+task copyToLib(type: Copy) {
+    into "$project.projectDir/lib"
+    from configurations.externalJars
+}
+
+def moduleName = "config"
+def moduleOrg = "ballerinax"
+def ballerinaConfigFile = new File("$project.projectDir/Ballerina.toml")
+def artifactBallerinaDocs = file("$project.projectDir/build/docs_parent/")
+def artifactCacheParent = file("$project.projectDir/build/cache_parent/")
+def artifactLibParent = file("$project.projectDir/build/lib_parent/")
+def tomlVersion = project.version.split("-")[0]
+def ballerinaCentralAccessToken = System.getenv('BALLERINA_CENTRAL_ACCESS_TOKEN')
+def targetBallerinaJar = file("$project.projectDir/target/caches/jar_cache/${moduleOrg}/${moduleName}/${tomlVersion}/${moduleOrg}-${moduleName}-${tomlVersion}.jar")
+def targetNativeJar = file("$project.rootDir/${moduleName}-native/build/libs/${moduleName}-native-${project.version}.jar")
+
+task ballerinaBuild {
+    inputs.dir file(project.projectDir)
+    doLast {
+        def originalConfig = ballerinaConfigFile.text
+        def newConfig = ballerinaConfigFile.text.replace("@project.version@", project.version)
+        newConfig = newConfig.replace("@toml.version@", tomlVersion)
+        ballerinaConfigFile.text = newConfig
+
+        def distributionBinPath =  project.projectDir.absolutePath + "/build/target/extracted-distributions/jballerina-tools-zip/jballerina-tools-${ballerinaLangVersion}/bin"
+
+        // Build and populate caches
+        exec {
+            workingDir project.projectDir
+            environment "JAVA_OPTS", "-DBALLERINA_DEV_COMPILE_BALLERINA_ORG=true"
+            if (Os.isFamily(Os.FAMILY_WINDOWS)) {
+                commandLine 'cmd', '/c', "$distributionBinPath/ballerina.bat build -c ${moduleName}"
+            } else {
+                commandLine 'sh', '-c', "$distributionBinPath/ballerina build -c ${moduleName}"
+            }
+        }
+        copy {
+            from file("$project.projectDir/Ballerina.toml")
+            into file("$artifactCacheParent/caches/bir/${moduleOrg}/${moduleName}/${tomlVersion}")
+        }
+        copy {
+            from file("$project.projectDir/Ballerina.toml")
+            into file("$artifactCacheParent/caches/balo/${moduleOrg}/${moduleName}/${tomlVersion}")
+        }
+        copy {
+            from file("$project.projectDir/Ballerina.toml")
+            into file("$artifactCacheParent/caches/jar/${moduleOrg}/${moduleName}/${tomlVersion}")
+        }
+        copy {
+            from file("$project.projectDir/target/caches/bir_cache")
+            into file("$artifactCacheParent/caches/bir")
+        }
+        copy {
+            from file("$project.projectDir/target/balo")
+            into file("$artifactCacheParent/caches/balo/${moduleOrg}/${moduleName}/${tomlVersion}")
+        }
+        copy {
+            from targetBallerinaJar
+            into file("$artifactCacheParent/caches/jar/${moduleOrg}/${moduleName}/${tomlVersion}")
+        }
+
+        copy {
+            from targetBallerinaJar
+            into file("$artifactLibParent/libs")
+        }
+
+        copy {
+            from targetNativeJar
+            into file("$artifactLibParent/libs")
+        }
+
+        copy {
+            from configurations.externalJars
+            into file("$artifactLibParent/libs")
+        }
+
+        // Publish to central
+        if (!project.version.endsWith('-SNAPSHOT') && ballerinaCentralAccessToken != null && project.hasProperty("publishToCentral")) {
+            println("Publishing to the ballerina central..")
+            exec {
+                workingDir project.projectDir
+                environment "JAVA_OPTS", "-DBALLERINA_DEV_COMPILE_BALLERINA_ORG=true"
+                if (Os.isFamily(Os.FAMILY_WINDOWS)) {
+                    commandLine 'cmd', '/c', "$distributionBinPath/ballerina.bat push ${moduleName}"
+                } else {
+                    commandLine 'sh', '-c', "$distributionBinPath/ballerina push ${moduleName}"
+                }
+            }
+
+        }
+        // Doc creation and packing
+        exec {
+            workingDir project.projectDir
+            environment "JAVA_OPTS", "-DBALLERINA_DEV_COMPILE_BALLERINA_ORG=true"
+            if (Os.isFamily(Os.FAMILY_WINDOWS)) {
+                commandLine 'cmd', '/c', "$distributionBinPath/ballerina.bat doc ${moduleName}"
+            } else {
+                commandLine 'sh', '-c', "$distributionBinPath/ballerina doc ${moduleName}"
+            }
+        }
+        copy {
+            from file("$project.projectDir/target/apidocs/${moduleName}")
+            into file("$project.projectDir/build/docs_parent/docs/${moduleName}")
+        }
+        ballerinaConfigFile.text = originalConfig
+    }
+
+    outputs.dir artifactCacheParent
+    outputs.dir artifactBallerinaDocs
+    outputs.dir artifactLibParent
+}
+
+task createArtifactZip(type: Zip) {
+    destinationDirectory = file("$buildDir/distributions")
+    from ballerinaBuild
+}
+
+publishing {
+    publications {
+        mavenJava(MavenPublication) {
+            artifact source: createArtifactZip, extension: 'zip'
+        }
+    }
+
+    repositories {
+        maven {
+            name = "GitHubPackages"
+            url = uri("https://maven.pkg.github.com/ballerina-platform/module-ballerina-config")
+            credentials {
+                username = System.getenv("GITHUB_ACTOR")
+                password = System.getenv("GITHUB_TOKEN")
+            }
+        }
+    }
+}
+
+test {
+    environment "testEnvironmentVariable", "EnvironmentVariable"
+    environment "user_name", "b7auser"
+    environment "password", "b7apw"
+    environment "hello_http_host", "192.168.1.11"
+    environment "hello_http_port", "5656"
+    environment "hello_cache_enabled", "true"
+    environment "hello_eviction_fac", "0.2333333"
+    environment "helloneg_http_port", "invalid port"
+    environment "helloneg_eviction_fac", "invalid eviction factor"
+}
+
+ballerinaBuild.dependsOn ":config-native:build"
+unpackJballerinaTools.dependsOn copyToLib
+ballerinaBuild.dependsOn unpackJballerinaTools
+build.dependsOn ballerinaBuild
