+++ config-native/src/test/java/org/ballerinalang/stdlib/config/test/ConfigProcessorTest.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.stdlib.configtest;
+package org.ballerinalang.stdlib.config.test;
 
 import org.ballerinalang.bcl.parser.BConfig;
 import org.ballerinalang.config.ConfigProcessor;
