+++ config-native/src/test/java/org/ballerinalang/stdlib/config/test/util/JBallerinaTestInitializer.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.stdlib.configtest.util;
+package org.ballerinalang.stdlib.config.test.util;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
