+++ config-native/src/test/resources/datafiles/array_negative.conf
@@ -1,0 +1,2 @@
+[array]
+negative=[ 1, 2.0 ]
