+++ config-native/src/test/resources/datafiles/array_negative2.conf
@@ -1,0 +1,8 @@
+[[fruit]]
+  name = "apple"
+
+  [[fruit.variety]]
+    name = "red delicious"
+
+  [fruit.variety]
+    name = "granny smith"
