+++ config-native/src/test/resources/datafiles/complex-key-ballerina.conf
@@ -1,0 +1,17 @@
+["hello:sayHello".test.test1]
+"groups"="xxx,yyy"
+
+["hello:sayHello1:123"]
+groups="xxx,yyy,abc"
+
+[hello.test."xxx"]
+"groups.123"="zzzz"
+
+[a.b]
+"123/pqr/tz"="abc123"
+
+["a/b/r"]
+"456/pqr"="lmn789"
+
+[abc123]
+testKey="testVal"
