+++ config-native/src/test/resources/datafiles/default/ballerina.conf
@@ -1,0 +1,80 @@
+#
+# Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+#
+# WSO2 Inc. licenses this file to you under the Apache License,
+# Version 2.0 (the "License"); you may not use this file except
+# in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+
+"ballerina.http.host"="10.100.1.205"
+
+"ballerina.http.instances"="http1,http2,http3"
+"ballerina.env.Path"="${env:PATH}"
+"ballerina.log.format"="{{timestamp}}[yyyy-MM-dd HH:mm:ss,SSS] {{level}} [{{package}}:{{unit}}] [{{file}}:{{line}}] [{{worker}}] - \"{{msg}}\" {{err}}"
+"path.test"='C:\Users\John'
+expirationPeriod=1.8e+06
+
+[http1]
+port=8085
+caching_enabled=true
+eviction_factor=0.3455
+request_limit=50
+
+[http2]
+port=8081
+#host= http://$sysv.host
+
+[http3]
+port  =   7070
+
+[multiline.strings]
+txt="""Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore \
+et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip \
+ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore \
+eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt \
+mollit anim id est laborum."""
+
+txt2='''
+Hello
+World!!!
+'''
+
+[listenerConfig.keyStore]
+paths=["/etc", "/tmp", "/usr/lib/"]
+ports=[9090, 9091, 9092]
+
+[employee]
+name="John"
+city="Sydney"
+
+[array]
+multidimensional=[ [ 1, 2 ], ["a", "b", "c"] ]
+
+[[foo]]
+a=1
+b=2
+
+[[foo]]
+a=3
+b=4
+
+[[products]]
+name = "Hammer"
+sku = 738594937
+
+[[products]]
+
+[[products]]
+name = "Nail"
+sku = 284758393
+color = "gray"
