+++ config-native/src/test/resources/datafiles/encrypted-configs.conf
@@ -1,0 +1,4 @@
+"user.password"="@encrypted:{UtD9d+o6eHpqFnBxtvhb+RWXey7qm7xLMt6+6mrt9w0=}"
+"int.value"="@encrypted:{vr1bbDIctz4ybPTlEEnEyuKJ+4OP2NDGsPSCQYEnyi0=}"
+"float.value"="@encrypted:{bSFxiiu7sSsMaGCHOy86aTQ5XISqxRtrK7lfsAIa0WA=}"
+"boolean.value"="@encrypted:{o5rtwoLh7MxGvqTCZ6nbxp3Y2irNIUCsClb9V00Zb4w=}"
