+++ config-native/src/test/resources/datafiles/invalid-ballerina.conf
@@ -1,0 +1,32 @@
+#
+# Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+#
+# WSO2 Inc. licenses this file to you under the Apache License,
+# Version 2.0 (the "License"); you may not use this file except
+# in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+
+ballerina.http.host="10.100.1.200"
+
+ballerina.http.instances="http1,http2"
+ballerina.env.PATH="${env:PATH}"
+ballerina.log.$format="{{timestamp}}[yyyy-MM-dd HH:mm:ss,SSS] {{level}} [{{package}}:{{unit}}] [{{file}}:{{line}}] [{{worker}}] - \"{{msg}}\" {{err}}"
+ballerina.home="$sys{ballerina.home}"
+
+invalid.entry
+
+[http1]
+ballerina.http.port=8080
+
+[http2]
+ballerina.http.port=8081
