+++ config-native/src/test/resources/datafiles/invalid-secret.txt
@@ -1,0 +1,1 @@
+wrong secret
