+++ config-native/src/test/resources/datafiles/lookupenv.conf
@@ -1,0 +1,13 @@
+[hello]
+  [hello.http]
+    host="localhost"
+    port=9090
+  [hello.cache]
+    enabled=false
+  [hello.eviction]
+    fac=0.25
+  [hello.keyStore]
+    paths=["/etc","/tmp","/usr/lib/"]
+[database]
+"server.host"="192.168.1.1"
+"server.port"=3306
