+++ config-native/src/test/resources/datafiles/lookupenv_negative.conf
@@ -1,0 +1,2 @@
+[helloneg.http]
+  port=8080
