+++ config-native/src/test/resources/datafiles/lookupenv_negative2.conf
@@ -1,0 +1,2 @@
+[helloneg.eviction]
+  fac=0.25
