+++ config-native/src/test/resources/datafiles/lookupenv_negative3.conf
@@ -1,0 +1,5 @@
+[helloneg]
+[helloneg.http]
+port=8080
+[helloneg.eviction]
+fac=0.25
