+++ config-native/src/test/resources/datafiles/secret.txt
@@ -1,0 +1,1 @@
+ballerina
