+++ config-native/src/test/resources/datafiles/service-config.conf
@@ -1,0 +1,4 @@
+backendEP.port="7070"
+
+[hello]
+basePath="/hello"
