+++ gradle.properties
@@ -1,0 +1,5 @@
+org.gradle.caching=true
+group=org.ballerinalang
+version=1.0.0-SNAPSHOT
+ballerinaLangVersion=2.0.0-Preview3
+ballerinaTomlParserVersion=1.2.2
